Arreglo empaquetado (conda test)
Se arregla el empaquetado. Se probó con bash (cmd) para cargar y se limpio tanto el archivo principal como bandas y tutorial. Probé la funcionalidad del paquete.
