+++ src/opt_pyFE/tutorial_optyFE.py
@@ -5,16 +5,13 @@ Created on Tue May 13 13:58:35 2025
 @author: david
 """
 #importamos el paquete de optimización de portafolios
-import opt_pyFE
+import opt_pyFE as opt
+import yfinance as yf
+import pandas as pd
+import numpy as np
+from sklearn.linear_model import LinearRegression
+import matplotlib.pyplot as plt
 
-file_1 = open('opt_pyFE/opt_pyFE_estandarizado.py').read()
-exec(file_1)
-
-file_2 = open('opt_pyFE/defmodelmach.py').read()
-exec(file_2)
-
-file_3 = open('opt_pyFE/Empaquetado_bandas_bollinger_analiticos.py').read()
-exec(file_3)
 
 #Definimos los parametros con los que trabajaremos y las especificaciones
 #de nuestro analisis 
@@ -46,40 +43,40 @@ end_date = "2025-05-26"
 #Generamos un cilo for para graficar la media movil y la tendencia de cada
 #uno de nuestros tickers 
 for ticker in tickers:
-    proyeccion(tickers)
+    opt.proyeccion(tickers, start_date, end_date)
 
 #La media movil se calcula por defecto a 50 días pero es posible ajustar
 #este parametro especificando el valor de window 
 for ticker in tickers:
-    proyeccion(tickers, window = 20)
+    opt.proyeccion(tickers, start_date, end_date, window = 20)
     
 #Se genera un grafico de bandas de bollinger por cada una de los tickers
-bandas_bollinger(tickers, start_date, end_date)
+opt.bandas_bollinger(tickers, start_date, end_date)
 
 
 #MODULO DE OPTIMIZACION#
 
 #Se descargan y tratan los datos para simular el peso de los portafolios
-df_aj = descargar_datos(tickers, start_date, end_date)
+df_aj = opt.descargar_datos(tickers, start_date, end_date) 
 
 #se calcula el rendimiento logaritmico de las emisoras
-log_returns = calcular_rendimientos_log(df_aj)
+log_returns = opt.calcular_rendimientos_log(df_aj) ## @David esto crea Warning revisar
 
 #Se simula un numero definido de portafolios para poder encontrar el optimo
 num_portafolios = 10000  
-weight, ReturnEsp, VolEsp, RadioSharpe = simular_portafolios(log_returns, 
+weight, ReturnEsp, VolEsp, RadioSharpe = opt.simular_portafolios(log_returns, 
                                                              num_portafolios)
 
 #Se encuentra el portofalio que miniza el riesgo entre las simulaciones
 #tomando esta como la distribución de pesos optima
-best_weights, retorno, volatilidad, sharpe_ratio = encontrar_mejor_portafolio(weight, 
+best_weights, retorno, volatilidad, sharpe_ratio = opt.encontrar_mejor_portafolio(weight, 
                                                                               ReturnEsp, 
                                                                               VolEsp, 
                                                                               RadioSharpe)
 
 #Se imprime en la consola un resumen de los resultados obtenidos y
 #de los pesos seleccionados
-mostrar_resultados(tickers, 
+opt.mostrar_resultados(tickers, 
                    best_weights, 
                    retorno, 
                    volatilidad, 
@@ -89,7 +86,7 @@ mostrar_resultados(tickers,
 
 #Calculamos los rendimientos y geenramos una matriz de covarianza 
 #para poder calcular los riesgo de nuestro portafolio
-rendimiento, rendimiento_medio, covmatrix = getdata(tickers, 
+rendimiento, rendimiento_medio, covmatrix = opt.getdata(tickers, 
                                                     start_date, 
                                                     end_date)
 #retiramos los valores nulos de nuestro datafreame 
@@ -107,17 +104,17 @@ inversion_inicial = 100000 #Definimos el monto a invertir
 rendimiento['portafolio'] = rendimiento.dot(peso) #dot calcula el producto 
 
 #Calculamos el rendimiento esperado y la desviacion estandar anualizada 
-pRet, pStd = desempeno(peso, 
+pRet, pStd = opt.desempeno(peso, 
                        rendimiento_medio, 
                        covmatrix, 
                        time)
 
 #Con los parametros obtenidos de los datos historicos calculamos el Var
-hVaR = -historicalVar(rendimiento['portafolio'], alpha=5)*np.sqrt(time)
+hVaR = -opt.historicalVar(rendimiento['portafolio'], alpha=5)*np.sqrt(time)
 print(' historical VaR 95th CI   :    ', round(inversion_inicial*hVaR, 2))
 
 #Con los parametros obtenidos de los datos historicos calculamos el Var condicional
-hCVaR = -historicalCVar(rendimiento['portafolio'], alpha=5)*np.sqrt(time)
+hCVaR = -opt.historicalCVar(rendimiento['portafolio'], alpha=5)*np.sqrt(time)
 print(' historical CVaR 95th CI  :    ', round(inversion_inicial*hCVaR, 2))
 
 
@@ -125,19 +122,19 @@ mc_sims = 1000 # numero de simulaciones
 T = 100 #periodo de tiempo en dias 
 
 #Se generan las simulaciones montecarlo 
-portfolio_sims, portResults = MonteCarlo(mc_sims, 
+portfolio_sims, portResults = opt.MonteCarlo(mc_sims, 
                                          T, 
                                          rendimiento_medio, 
                                          peso, inversion_inicial, 
                                          covmatrix)
 
 #Con los parametros obtenidos de los datos historicos calculamos el Var
-MCVaR = inversion_inicial - mcVaR(portResults, alpha=5)
+MCVaR = inversion_inicial - opt.mcVaR(portResults, alpha=5)
 print(" MC VaR  95th CI          :    ", round(MCVaR, 2))
 
 #Con los parametros obtenidos de los datos historicos calculamos el Var condicional
-MCCVaR = inversion_inicial - mcCVaR(portResults, alpha=5)
+MCCVaR = inversion_inicial - opt.mcCVaR(portResults, alpha=5)
 print(" MC CVaR 95th CI          :    ", round(MCCVaR, 2))
 
 #Se imprime un resumen de los valores de todos los valores calculados
-resum(inversion_inicial, inversion_inicial, hVaR, hCVaR, MCVaR, MCCVaR, pRet)
+opt.resum(inversion_inicial, inversion_inicial, hVaR, hCVaR, MCVaR, MCCVaR, pRet)
